+++ src/Opulence/Bootstrappers/BootstrapperRegistry.php
@@ -9,7 +9,6 @@
 namespace Opulence\Bootstrappers;
 
 use Opulence\Applications\Environments\Environment;
-use Opulence\Applications\Paths;
 use RuntimeException;
 
 /**