+++ src/Opulence/Framework/Http/Middleware/Session.php
@@ -12,7 +12,7 @@
 namespace Opulence\Framework\Http\Middleware;
 
 use Closure;
-use Opulence\Applications\Paths;
+use Opulence\Bootstrappers\Paths;
 use Opulence\Http\Middleware\IMiddleware;
 use Opulence\Http\Requests\Request;
 use Opulence\Http\Responses\Response;